+++ gate/src/main/java/org/wangyefeng/game/gate/protocol/Protocol.java
@@ -2,5 +2,5 @@ package org.wangyefeng.game.gate.protocol;
 
 public interface Protocol {
 
-    int getCode();
+    short getCode();
 }
