+++ logic/src/main/java/org/wangyefeng/game/logic/protocol/Protocol.java
@@ -2,5 +2,5 @@ package org.wangyefeng.game.logic.protocol;
 
 public interface Protocol {
 
-    int getCode();
+    short getCode();
 }
